Try pinging the daemon instead of inspecting /proc.
